pyproject.toml
[tool.poetry]
name = "spy_bot"
version = "0.1.0"
description = ""
authors = ["levmorgot <levmorgot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.16"
asyncio = "^3.4.3"
kafka-python = "^2.0.2"
python-dotenv = "^0.19.2"
mysql-connector = "^2.2.9"
aiokafka = "^0.7.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
